Add <M-l> support for explicitly sized delimiters
Summary:
Writing `(<M-l>` in insert mode produces `\left(<++>\right)<++>`. But I
don't like `\left` and `\right`. This patch adds a flag that allows the
mapping to prompt for an explicit size, inserting, for instance,
`\Bigl(<++>\Bigr)<++>`. Leaving the size blank will revert to the
original behavior, and the flag is off by default.

Test Plan:
In insert mode, write `(` and press `<M-l>`. By default, the behavior
should be unchanged: the output should be

    \left(<++>\right)<++>

as before. Then write

    :let g:tex_use_sized_delimiters = 1

and try it again. Pressing "Enter" at the prompt should yield

    \Bigl(<++>\Bigr)<++>

while changing the size to `bigg` should have the expected result, and
removing the size entirely should fall back to

    \left(<++>\right)<++>

as before.