Add `o ↦ \omega, `O ↦ \Omega bindings
Summary:
The bindings for tick-`o` and tick-`O` were unused, presumably because
omicron doesn't have a TeX command (which makes sense). But we might as
well bind to omega, instead!

Test Plan:
In a TeX file, type

    2 `p f = `o

and see that it renders as

    2 \pi f = \omega

as desired.

I also ran `git grep kappa` to make sure that I wasn't missing any
files that needed to be changed. There are two occurrences that I didn't
change: they are `mathmacros.vim` and `mathmacros-utf.vim`. Each of
these already has an omega command, so no patch was required.